+++ src/Test.CoreLib/src/System/Runtime/CompilerServices/EagerStaticClassConstructionAttribute.cs
@@ -1,0 +1,13 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime.CompilerServices
+{
+    // When applied to a type this custom attribute will cause any static class constructor to be run eagerly
+    // at module load time rather than deferred till just before the class is used.
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct, Inherited = false, AllowMultiple = false)]
+    public class EagerStaticClassConstructionAttribute : Attribute
+    {
+    }
+}
