--- src/Test.CoreLib/src/System/RuntimeTypeHandle.Ext.cs
@@ -1,17 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Runtime.CompilerServices;
-
-namespace System
-{
-    partial struct RuntimeTypeHandle
-    {
-        [Intrinsic]
-        internal unsafe static IntPtr GetValueInternal(RuntimeTypeHandle handle)
-        {
-            return (IntPtr)handle._pEEType.ToPointer();
-        }
-    }
-}
